add firebase_analytics
